fix: build apps using the productive environment
We have noticed that on GitHub Actions, Parcel builds the host app
using the local environment config (.env) instead of the production one
(.env.production). When building the app locally, not on GitHub, it
works as expected. Actually, there should be no need to specify
`NODE_ENV=development` or `NODE_ENV=production` since that
is implied by Parcel's serve/watch or build, respectively.