chore: migrate to the latest version of Parcel to fix build errors
By default, Parcel does not build a web application if it (or a dependency) depends on Node modules. Specifically, it is the `js-sha256` library used by the SCION Microfrontend Platform that checks for the presence of the `process` Node module. Since this is only a check, we instruct Parcel to replace the `process` Node module with an empty module by setting a "Ship Alias" in the `package.json`. For more information, refer to https://parceljs.org/features/dependency-resolution/#shim-aliases.